Add files via upload
Historical Wildfires

This file is part of the practice project of Week 5 of the certificate "Data Visualisation with Python" which is part of the IBM Data Science Professional Certificate offered via Coursera. The file includes the dataset as well as the analysis done which is part of the exercises. The file examines various questions around the wildfires in Australia,